chore(lint): apply what the RUF rules found
The newly enabled RUF family surfaced 120 findings; all are resolved. The
bulk is mechanical: 49 unsorted __all__ lists, 16 dead noqa comments, 17
unused unpacked variables in tests, and the import-order fallout. The
by-hand tail: pytest.raises match= patterns become raw strings (the
intentional regexes) or get their literal dots escaped; six successive-pair
zips become itertools.pairwise; four literal concatenations become
unpacking; two redundant int() casts are dropped. RUF001-003
(ambiguous-unicode) are ignored with a comment: the docstrings here use
unicode maths characters on purpose. ruff format now passes tree-wide (it
gates in CI as of the previous commit), and the fast tier is green: 988
passed, 10 skipped in 66 s.

+++ libs/qscat/qscat/tuning/metrics.py
@@ -73,7 +73,7 @@ def tensor_cost(g_r: FemDvrEcsGrid, g_R: FemDvrEcsGrid) -> dict[str, Any]:
     footprint.
     """
     n_unknowns = g_r.n * g_R.n
-    est_nnz = int(round(n_unknowns * _NNZ_PER_ROW))
+    est_nnz = round(n_unknowns * _NNZ_PER_ROW)
     est_factor_gib = _GIB_COEFF * float(n_unknowns) ** _GIB_EXPONENT
     est_factor_seconds = _SECONDS_COEFF * float(n_unknowns) ** _SECONDS_EXPONENT
     return {
